TLDR Refer to FEL and NESTCOMP for additional component serializing
In addition to the previous message on component serialization, if you have struct/class variable, you would need to have two custom friend functions.

There are many ways to (de)serialize but prefably use the format of FEL and NESTCOMP.

FEL: Example of REGISTERED Component, will be yeeted and deleted at the end of M1

NESTCOMP: Example of a nested struct, can be a possible registered component or a simple struct. Probably be used for custom Math struct or wrapping other lib structs